+++ common/src/main/java/net/witcher_rpg/blocks/WitcherBlocks.java
@@ -1,7 +1,7 @@
 package net.witcher_rpg.blocks;
 
-import net.fabricmc.fabric.api.itemgroup.v1.ItemGroupEvents;
-import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
+import net.spell_engine.PlatformEvents;
+import net.minecraft.block.AbstractBlock;
 import net.minecraft.block.Block;
 import net.minecraft.block.Blocks;
 import net.minecraft.block.ExperienceDroppingBlock;
@@ -35,49 +35,49 @@ public class WitcherBlocks {
     }
 
     public static final Entry SILVER_ORE = entry("silver_ore", "Silver Ore",new ExperienceDroppingBlock(UniformIntProvider.create(3, 7),
-            FabricBlockSettings.create()
+            AbstractBlock.Settings.create()
                     .mapColor(MapColor.STONE_GRAY)
                     .instrument(NoteBlockInstrument.BASEDRUM)
                     .requiresTool()
                     .strength(3.0F, 3.0F)
     ));
     public static final Entry DEEPSLATE_SILVER_ORE= entry("deepslate_silver_ore", "Deepslate Silver Ore",new ExperienceDroppingBlock(UniformIntProvider.create(3, 7),
-            FabricBlockSettings.create()
+            AbstractBlock.Settings.create()
                     .mapColor(MapColor.STONE_GRAY)
                     .instrument(NoteBlockInstrument.BASEDRUM)
                     .requiresTool()
                     .strength(4.0F, 3.0F)
     ));
     public static final Entry METEORITE_ORE= entry("meteorite_ore", "Meteorite Ore",new ExperienceDroppingBlock(UniformIntProvider.create(3, 7),
-            FabricBlockSettings.create()
+            AbstractBlock.Settings.create()
                     .mapColor(MapColor.TERRACOTTA_BROWN)
                     .instrument(NoteBlockInstrument.BASEDRUM)
                     .requiresTool()
                     .strength(4.0F, 3.0F)
     ));
     public static final Entry DEEPSLATE_DARK_IRON_ORE= entry("deepslate_dark_iron_ore", "Deepslate Dark Iron Ore",new ExperienceDroppingBlock(UniformIntProvider.create(3, 7),
-            FabricBlockSettings.create()
+            AbstractBlock.Settings.create()
                     .mapColor(MapColor.STONE_GRAY)
                     .instrument(NoteBlockInstrument.BASEDRUM)
                     .requiresTool()
                     .strength(4.0F, 3.0F)
     ));
     public static final Entry DARK_IRON_ORE= entry("dark_iron_ore", "Dark Iron Ore",new ExperienceDroppingBlock(UniformIntProvider.create(3, 7),
-            FabricBlockSettings.create()
+            AbstractBlock.Settings.create()
                     .mapColor(MapColor.STONE_GRAY)
                     .instrument(NoteBlockInstrument.BASEDRUM)
                     .requiresTool()
                     .strength(3.0F, 3.0F)
     ));
     public static final Entry NETHER_DARK_IRON_ORE= entry("nether_dark_iron_ore", "Nether Dark Iron Ore",new ExperienceDroppingBlock(UniformIntProvider.create(3, 7),
-            FabricBlockSettings.create()
+            AbstractBlock.Settings.create()
                     .mapColor(MapColor.DARK_RED)
                     .instrument(NoteBlockInstrument.BASEDRUM)
                     .requiresTool()
                     .strength(3.0F, 3.0F)
     ));
     public static final Entry SILVER_BLOCK = entry("silver_block", "Block of Silver",new Block(
-            FabricBlockSettings.create()
+            AbstractBlock.Settings.create()
                     .mapColor(MapColor.LIGHT_BLUE_GRAY)
                     .instrument(NoteBlockInstrument.IRON_XYLOPHONE)
                     .requiresTool()
@@ -85,14 +85,14 @@ public class WitcherBlocks {
                     .sounds(BlockSoundGroup.METAL)
     ));
     public static final Entry RAW_SILVER_BLOCK = entry("raw_silver_block", "Block of Silver Ore",new Block(
-            FabricBlockSettings.create()
+            AbstractBlock.Settings.create()
                     .mapColor(MapColor.LIGHT_GRAY)
                     .instrument(NoteBlockInstrument.BASEDRUM)
                     .requiresTool()
                     .strength(5.0F, 6.0F)
     ));
     public static final Entry STEEL_BLOCK = entry("steel_block", "Block of Steel",new Block(
-            FabricBlockSettings.create()
+            AbstractBlock.Settings.create()
                     .mapColor(MapColor.STONE_GRAY)
                     .instrument(NoteBlockInstrument.IRON_XYLOPHONE)
                     .requiresTool()
@@ -100,7 +100,7 @@ public class WitcherBlocks {
                     .sounds(BlockSoundGroup.METAL)
     ));
     public static final Entry METEORITE_BLOCK = entry("meteorite_block", "Block of Meteorite",new Block(
-            FabricBlockSettings.create()
+            AbstractBlock.Settings.create()
                     .mapColor(MapColor.PALE_PURPLE)
                     .instrument(NoteBlockInstrument.IRON_XYLOPHONE)
                     .requiresTool()
@@ -108,7 +108,7 @@ public class WitcherBlocks {
                     .sounds(BlockSoundGroup.METAL)
     ));
     public static final Entry DARK_IRON_BLOCK = entry("dark_iron_block", "Block of Dark Iron",new Block(
-            FabricBlockSettings.create()
+            AbstractBlock.Settings.create()
                     .mapColor(MapColor.DARK_GREEN)
                     .instrument(NoteBlockInstrument.IRON_XYLOPHONE)
                     .requiresTool()
@@ -116,7 +116,7 @@ public class WitcherBlocks {
                     .sounds(BlockSoundGroup.METAL)
     ));
     public static final Entry METEORITE_SILVER_BLOCK = entry("meteorite_silver_block","Block of Meteorite Silver", new Block(
-            FabricBlockSettings.create()
+            AbstractBlock.Settings.create()
                     .mapColor(MapColor.PALE_PURPLE)
                     .instrument(NoteBlockInstrument.IRON_XYLOPHONE)
                     .requiresTool()
@@ -124,7 +124,7 @@ public class WitcherBlocks {
                     .sounds(BlockSoundGroup.NETHERITE)
     ));
     public static final Entry DARK_STEEL_BLOCK = entry("dark_steel_block", "Block of Dark Steel",new Block(
-            FabricBlockSettings.create()
+            AbstractBlock.Settings.create()
                     .mapColor(MapColor.PALE_GREEN)
                     .instrument(NoteBlockInstrument.IRON_XYLOPHONE)
                     .requiresTool()
@@ -138,7 +138,7 @@ public class WitcherBlocks {
             Registry.register(Registries.BLOCK, Identifier.of(WitcherClassMod.MOD_ID, entry.name), entry.block);
             Registry.register(Registries.ITEM, Identifier.of(WitcherClassMod.MOD_ID, entry.name), entry.item());
         }
-        ItemGroupEvents.modifyEntriesEvent(WitcherGroup.WITCHER_KEY).register((content) -> {
+        PlatformEvents.onItemGroupModify(WitcherGroup.WITCHER_KEY, (content, context) -> {
             for (var entry : all) {
                 content.add(entry.item());
             }