+++ common/src/main/java/net/witcher_rpg/entity/WitcherEntities.java
@@ -1,7 +1,6 @@
 package net.witcher_rpg.entity;
 
-import net.fabricmc.fabric.api.object.builder.v1.entity.FabricEntityTypeBuilder;
-import net.minecraft.entity.EntityDimensions;
+import net.minecraft.entity.EntityType;
 import net.minecraft.entity.SpawnGroup;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.Registry;
@@ -14,22 +13,22 @@ public class WitcherEntities {
         YrdenEntity.ENTITY_TYPE = Registry.register(
                 Registries.ENTITY_TYPE,
                 Identifier.of(MOD_ID, "yrden"),
-                FabricEntityTypeBuilder.<YrdenEntity>create(SpawnGroup.MISC, YrdenEntity::new)
-                        .dimensions(EntityDimensions.changing(6F, 0.5F))
-                        .fireImmune()
-                        .trackRangeBlocks(128)
-                        .trackedUpdateRate(20)
-                        .build()
+                EntityType.Builder.<YrdenEntity>create(YrdenEntity::new, SpawnGroup.MISC)
+                        .dimensions(6F, 0.5F)
+                        .makeFireImmune()
+                        .maxTrackingRange(128)
+                        .trackingTickInterval(20)
+                        .build("yrden")
         );
         YrdenMagicTrapEntity.ENTITY_TYPE = Registry.register(
                 Registries.ENTITY_TYPE,
                 Identifier.of(MOD_ID, "yrden_magical_trap"),
-                FabricEntityTypeBuilder.<YrdenMagicTrapEntity>create(SpawnGroup.MISC, YrdenMagicTrapEntity::new)
-                        .dimensions(EntityDimensions.changing(6F, 0.5F))
-                        .fireImmune()
-                        .trackRangeBlocks(128)
-                        .trackedUpdateRate(20)
-                        .build()
+                EntityType.Builder.<YrdenMagicTrapEntity>create(YrdenMagicTrapEntity::new, SpawnGroup.MISC)
+                        .dimensions(6F, 0.5F)
+                        .makeFireImmune()
+                        .maxTrackingRange(128)
+                        .trackingTickInterval(20)
+                        .build("yrden_magical_trap")
         );
     }
 }